LayoutTests/storage/indexeddb/tutorial-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Please view source for more information on what this is doing and why...

Everything worked!
